+++ src/__tests__/products.test.ts
@@ -115,7 +115,7 @@ Object {
   "endpoint": "products/123",
   "method": "GET",
   "params": Object {
-    "with": "images.attributes:legacy(false),categories,categories.properties",
+    "with": "images.attributes:legacy(false),categories",
   },
 }
 `);
@@ -132,7 +132,7 @@ Object {
   "endpoint": "products/123",
   "method": "GET",
   "params": Object {
-    "with": "images.attributes:legacy(false),categories",
+    "with": "images.attributes:legacy(false),categories:properties()",
   },
 }
 `);